msm: mdss: unlock offlock if exiting ESD during shutdown
Previously, the device could hang while inside the shutdown notifier
mdss_fb_shutdown(). The hang occurred in mdss_mdp_cmd_stop() while
trying to lock the offlock designed to prevent race conditions between
ESD and powering off the display.

The solution is to unlock the offlock if ESD is exited early because
display driver shutdown is in progress.

This bug was caused by an interaction between these changes:
14a9091 (CR) msm: mdss: fix race condition of ESD status check
99b8b32 msm: mdss: Prevent race condition between dsi off and bta status check

Signed-off-by: Patrick Auchter <patrick.auchter@motorola.com>
Signed-off-by: Fred Fettinger <fred.fettinger@motorola.com>
Reviewed-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>
Reviewed-by: Liang-Tao Gao <w20048@motorola.com>
Reviewed-by: Joseph Swantek <jswantek@motorola.com>