msm: mdss: Reordering ULPS exiting seq
- Depending on the HWs, sometimes after exiting ULPS mode, some MIPI
lines are not in the STOPSTATE this will prevent the host to communicate
with panel or vice versa. It also prevents the host to put the MIPI lines
back to the ULPS mode.
- This change is in QCOM baseline. this change will cause another issue
that the ULPS escape seq which will violate the MIPI standard, but some
how the display will not able to get out from ULPS. From testing,
this fixes the entering ULPS mode issue and no other side effect of
this change.

- This patch is a workaround, and the new fix will address under
CR#(CR)

Signed-off-by: Bang Nguyen <bnguyen1@motorola.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>