Revert "mm: make vmstat_update periodic run conditional"
This reverts commit 7d252cd22a3f6cb459e8b012912dfd258157f7df because it
has been implicated in kernel crashes when its workqueue timer is
migrated during CPU hotplug.

When HMP was initially being developed against a 3.4 derived kernel, it
was observed that wakeups were occurring every 30s across every core to
give the vmstat accounting a kick. This was causing a noticeable
increase in energy consumption on the really quiet use cases such as
audio and video playback. So commit 7d252cd22a was used which turned off
the updates on idle CPUs to reduce that.

On the 3.10 derived LSK this revert does not result in a significant
increase in power consumption so it is assumed that changes since 3.4
have mitigated the initial problem.

Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Kevin Hilman <khilman@linaro.org>